% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kantorovich.R
\name{discrete}
\alias{discrete}
\title{Discrete 0-1 distance}
\usage{
discrete(x, y, gmp = FALSE)
}
\arguments{
\item{x, y}{vectors}

\item{gmp}{logical, use exact arithmetic with the help of the \code{gmp} package}
}
\description{
Returns the 0-1 distance between  \code{x} and \code{y}
}
\examples{
discrete(2, 3)
if(require(gmp)){
 discrete(as.bigq(2), as.bigq(3), gmp=TRUE)
}

}

