% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kantorovich.R
\name{edistances}
\alias{edistances}
\title{Extremal distances}
\usage{
edistances(mu, nu, dist = NULL, ...)
}
\arguments{
\item{mu}{row margins}

\item{nu}{column margins}

\item{dist}{function or matrix, the distance to be minimized on average. If \code{NULL}, the 0-1 distance is used.}

\item{...}{arguments passed to \code{dist}}
}
\value{
a list with two components: the extreme joinings in a list and the distances in a vector
}
\description{
Compute the distances at the extreme points
}

