\name{kantorovich-package}
\alias{kantorovich-package}
\docType{package}
\title{
Kantorovich distance between discrete probability measures
}
\description{
Computes the Kantorovich distance between two discrete probability measures.

To learn more, start with the vignettes:
\code{browseVignettes(package="kantorovich")}.

If you encounter a bug, or if you have a suggestion to improve the package, please file an issue on the github repo \url{https://github.com/stla/kantorovich}.
}
\details{
\tabular{ll}{
Package: \tab kantorovich\cr
Type: \tab Package\cr
Version: \tab 0.0.0\cr
Date: \tab 2016-01-10\cr
License: \tab GPL-2\cr
}
}

\author{
Stéphane Laurent
}
