% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kantorovich.R
\name{names.bigq}
\alias{names.bigq}
\title{Names for bigq}
\usage{
\method{names}{bigq}(x)
}
\arguments{
\item{x}{a \code{bigq} vector}
}
\value{
the names of \code{x}
}
\description{
attention car l'accès x["a"] ne marche pas
je ne m'en sers que dans arrange_names(), ce serait pas plus simple de convertir en character ?
}

