% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kantorovich.R
\name{Vectorize_bigq}
\alias{Vectorize_bigq}
\title{Vectorize a function returning a bigq number}
\usage{
Vectorize_bigq(f)
}
\arguments{
\item{f}{a function of two variables returing a bigq scalar}
}
\value{
the vectorized version of \code{f}
}
\description{
Vectorize a function returning a bigq number
}
\examples{
if(require(gmp)){
 f <- function(x,y){
   if(x==y) return(as.bigq(0)) else return(x+y)
 }
 x <- as.bigq(c(0,0)); y <- as.bigq(c(0,1))
 f(x,y)
 g <- Vectorize_bigq(f)
 g(x,y)
}
}

