% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kantorovich.R
\name{ejoinings}
\alias{ejoinings}
\title{Extreme joinings}
\usage{
ejoinings(mu, nu)
}
\arguments{
\item{mu}{row margins}

\item{nu}{column margins}
}
\value{
a list with the extreme joinings (matrices)
}
\description{
Return extreme joinings between \code{mu} and \code{nu}
}
\examples{
mu <- nu <- c(0.5, 0.5)
ejoinings(mu, nu)
if(require(gmp)){
 # use exact arithmetic
 library(gmp)
 mu <- nu <- as.bigq(c(0.5,0.5))
 ejoinings(mu, nu)
}

}

