\name{kantorovich-package}
\alias{kantorovich-package}
\docType{package}
\title{
Kantorovich Distance Between Probability Measures
}
\description{
Computes the Kantorovich distance between two probability measures on a finite set.

To learn more, start with the vignettes:
\code{browseVignettes(package="kantorovich")}.

If you encounter a bug, or if you have a suggestion to improve the package, please file an issue on the github repo \url{https://github.com/stla/kantorovich}.
}
\details{
\tabular{ll}{
Package: \tab kantorovich\cr
Type: \tab Package\cr
Version: \tab 1.1.0.9000\cr
Date: \tab 2016-02-04\cr
License: \tab GPL-2\cr
}
}

\author{
Stéphane Laurent
}
