% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kantorovich.R
\name{kantorovich}
\alias{kantorovich}
\title{Kantorovich distance}
\usage{
kantorovich(mu, nu, dist = NULL, ...)
}
\arguments{
\item{mu}{row margins}

\item{nu}{column margins}

\item{dist}{function, the distance to be minimized on average. If \code{NULL}, the 0-1 distance is used.}

\item{...}{arguments passed to \code{dist}}
}
\value{
the Kantorovich distance
}
\description{
Compute the Kantorovich distance
}
\examples{
mu <- c(1/7, 2/7, 4/7)
nu <- c(1/4, 1/4, 1/2)
kantorovich(mu, nu)
if(require(gmp)){
 library(gmp)
 mu <- as.bigq(c(1,2,4), 7)
 nu <- as.bigq(c(1,1,1), c(4,4,2))
 kantorovich(mu, nu)
}
}

