% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kantorovich.R
\name{names.bigq}
\alias{names.bigq}
\title{Names for bigq vectors}
\usage{
\method{names}{bigq}(x)
}
\arguments{
\item{x}{a \code{bigq} vector}
}
\value{
the names of \code{x}
}
\description{
Names for bigq vectors
}

